1、将draw源文件，可变栈声明更改为动态内存申请 2、屏蔽GUI文件非win32系统的printf函数 3、修正字体文件伪unicode汉字识别部分代码

+++ lv_font_chs_24.c
@@ -4257,7 +4257,7 @@ char *GetFakeUnicodeForCH(const char *pStr, int32_t s32StrLen)
 		while (pTmp < pEnd)
 		{
 			uint16_t u16Letter;
-			if (pTmp[0] > 0)
+			if ((int8_t)pTmp[0] > 0)
 			{
 				u16Letter = pTmp[0];
 				pTmp++;
